+++ src/main/java/io/webfolder/cdp/type/network/WebSocketResponse.java
@@ -1,5 +1,5 @@
 /**
- * cpd4j - Chrome DevTools Protocol for Java
+ * cdp4j - Chrome DevTools Protocol for Java
  * Copyright © 2017 WebFolder OÜ (support@webfolder.io)
  *
  * This program is free software: you can redistribute it and/or modify
@@ -26,7 +26,7 @@ import java.util.Map;
  */
 @Experimental
 public class WebSocketResponse {
-    private Double status;
+    private Integer status;
 
     private String statusText;
 
@@ -41,14 +41,14 @@ public class WebSocketResponse {
     /**
      * HTTP response status code.
      */
-    public Double getStatus() {
+    public Integer getStatus() {
         return status;
     }
 
     /**
      * HTTP response status code.
      */
-    public void setStatus(Double status) {
+    public void setStatus(Integer status) {
         this.status = status;
     }
